+++ src/forget_command.rs
@@ -20,9 +20,14 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
+use std::fs;
+use std::path::PathBuf;
+
 use crate::Command;
-use crate::process_garbage_collection;
+use crate::ERROR_LOADING_REPOSITORY_FAILED;
+use crate::ERROR_READING_DIRECTORY_FAILED;
 use crate::Repository;
+use crate::Revision;
 use crate::ZatsuError;
 
 pub struct ForgetCommand {
@@ -58,4 +63,109 @@ impl ForgetCommand {
 	    revision_count,
 	}
     }
+}
+
+fn process_garbage_collection() -> Result<(), ZatsuError> {
+    let repository = match Repository::load(".zatsu/repository.json") {
+        Ok(repository) => repository,
+        Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_LOADING_REPOSITORY_FAILED)),
+    };
+
+    let read_dir = match fs::read_dir(".zatsu/revisions") {
+        Ok(read_dir) => read_dir,
+        Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_READING_DIRECTORY_FAILED)),
+    };
+    let mut revision_paths: Vec<PathBuf> = Vec::new();
+    for result in read_dir {
+        if result.is_ok() {
+            let entry = result.unwrap();
+            revision_paths.push(entry.path());
+        }
+    }
+
+    for path in revision_paths {
+        let read_dir = match fs::read_dir(path) {
+            Ok(read_dir) => read_dir,
+            Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_READING_DIRECTORY_FAILED)),
+        };
+        for result in read_dir {
+            if result.is_ok() {
+                let entry = result.unwrap();
+                let path = entry.path();
+                let mut found = false;
+
+                let option = path.file_stem();
+                if option.is_some() {
+                    let file_stem = option.unwrap().to_string_lossy();
+                    println!("file_stem: {}", file_stem);
+                    let result = file_stem.parse();
+                    if result.is_ok() {
+                        let revision_number: i32 = result.unwrap();
+                        let option = repository.revision_numbers.iter().find(|&value| *value == revision_number);
+                        if option.is_some() {
+                            found = true;
+                        }
+                    }
+                }
+
+                if !found {
+                    match fs::remove_file(path) {
+                        Ok(()) => (),
+                        Err(_) => (),
+                    }
+                }
+            }
+        }
+    }
+
+    let read_dir = match fs::read_dir(".zatsu/objects") {
+        Ok(read_dir) => read_dir,
+        Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_READING_DIRECTORY_FAILED)),
+    }; 
+    let mut object_paths: Vec<PathBuf> = Vec::new();
+    for result in read_dir {
+        if result.is_ok() {
+            let entry = result.unwrap();
+            object_paths.push(entry.path());
+        }
+    }
+
+    for path in object_paths {
+        let read_dir = match fs::read_dir(path) {
+            Ok(read_dir) => read_dir,
+            Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_READING_DIRECTORY_FAILED)),
+        };
+
+        for result in read_dir {
+            if result.is_ok() {
+                let entry = result.unwrap();
+                let path = entry.path();
+                let option = path.file_name();
+                if option.is_some() {
+                    let hash = option.unwrap().to_string_lossy();
+                    let mut found = false;
+                    for revision_number in &repository.revision_numbers {
+                        let result = Revision::load(format!(".zatsu/revisions/{:02x}/{}.json", revision_number & 0xFF, revision_number));
+                        if result.is_ok() {
+                            let revision = result.unwrap();
+                            for entry in revision.entries {
+                                if entry.hash == hash {
+                                    found = true;
+                                }
+                            }
+                        }
+                    }
+
+                    if !found {
+                        match fs::remove_file(path) {
+                            Ok(()) => (),
+                            Err(_) => (),
+                        };
+                    }
+                }
+            }
+        }
+    }
+
+    Ok(())
 }
